Refactor to separate command handling from input
The idea is that we'll soon have a httpd server which will provide a
different way to receive commands, so if we split the command
handling logic out from the CLI and command input logic then we
should be able to reuse the command handler with the httpd server.

Also there's a "dummy" command handler which acts like the real one
but doesn't require access to the real hardware. To use it you need
to modify move_cli.py and check which of the two import lines are
used.

(Note that there could be bugs in the real command handler here
since I haven't been able to try on the real system yet.)