Allow cli to CPIO cleanup without importing GPIO
Define a static method in AlphaBot2 that we can call to run the GPIO
cleanup. This means we don't need to import GPIO in the CLI script,
which makes it easier to hack on this code when we're not on the
Raspberry Pi and the GPIO module is not available.

This is preparation for some refactoring of the way commands are
handled in an upcoming commit.