Use `finally` so cleanup happens on any exception
If we exit for some reason other than KeyboardInterrupt then
the cleanup won't run. Using "finally" should make sure it
runs no matter what.

And, because of that we can move back to using break to exit which I
think is better than raising KeyboardInterrupt when we want to do a
"soft" exit.

(Printing the extra newline on Ctrl-C makes it look nicer when you
exit that way.)